storage: support measuring without backing file
By introducing the new `withBacking` argument to Volume.measure it will
now be possible to measure internal volumes and getting their required
size only for them, and not the size of the entire chain including the
backing files.

This would relieve us of having to "guess" the required size in the
engine.

Measure on a file Storage Domain:
$ qemu-img measure --output json -O raw 'json:{"file": {"driver": "file", "filename": "/rhev/data-center/mnt/nfs:_root_storage__domains_sd1/bf5d4320-191d-4551-9c21-94745d1c6ec1/images/ef282a06-251b-4d17-a587-9e3337be6df2/0bae1526-f18f-4ede-98e8-2ec7c3e97f9e"}, "driver": "qcow2", "backing": null}'

Measure on a block Storage Domain:
$ qemu-img measure --output json -O raw 'json:{"file": {"driver": "host_device","filename": "/rhev/data-center/mnt/blockSD/9090d51a-fa55-478d-b77a-5ef07470ff3f/images/1d390146-6812-4405-9b40-ec7e7fe88d4d/4ee5265e-c0b5-4064-9ac5-b806fb3a0824"}, "driver":"raw"}'

Change-Id: I43b70f29284d89e6ec3ed41123535250fa049e9d
Bug-Url: https://bugzilla.redhat.com/1826365
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>